Make config/EvalConfig/io the inspect-free "contract" for solve environments
Splits agent-eval's inspect-bound scoring path from the inspect-free
"contract" (suite config schema, EvalConfig, git-reproducibility check,
atomic write) that solve environments import. Motivated by
allenai/gas2own#346: astabench PR #156 had to *copy* the suite-config
schema and verify_git_reproducibility into asta-bench because importing
them from agent-eval dragged in inspect-ai==0.3.203, which then wins uv's
overlay re-resolution and downgrades a solver env that needs a newer
inspect (e.g. inspect-swe). This lets those consumers import instead of
duplicate.

Changes (no behavior change on the scoring path):
- pyproject: move the inspect-ai==0.3.203 pin out of base dependencies
  into a `scoring` extra; `leaderboard` now includes `agent-eval[scoring]`.
  Scoring consumers install `agent-eval[scoring]` (or `[leaderboard]`).
- __init__.py: expose process_eval_logs / compute_summary_statistics /
  upload_folder_to_hf lazily (PEP 562 __getattr__) so importing agenteval
  or an inspect-free submodule no longer forces inspect.
- models.py: drop the module-level `from .score import TaskResult`
  (EvalConfig/SubmissionMetadata never needed it); move the only user,
  TaskResults, into score.py alongside TaskResult. models.py is now
  inspect-free.
- io.py: relocate verify_git_reproducibility here (was in cli.py, which
  imports inspect at module load) so solve envs can import it inspect-free.
- leaderboard/models.py: import SuiteConfig from ..config and TaskResult
  from ..score (no longer re-exported by ..models).

Verified: in a venv with no inspect_ai installed, agenteval,
agenteval.config, agenteval.models (EvalConfig schema resolves +
round-trips), and agenteval.io all import; the scoring path still
raises ModuleNotFoundError for inspect_ai as expected. black/isort/flake8
clean.
